[fix](function) Fix overflow in quarter arithmetic (#68410)
Problem Summary:
`quarters_add` / `quarters_sub` uses the i32 middle value when
converting quarters to months, and the oversized INT parameter may have
an integer backturn, bypassing the subsequent date range check and
returning the wrong date.

This repair includes:

1. **BE quarter addition and subtraction**: Use Int64 to receive the
quarterly offset and calculate the number of months to avoid overflow
when multiplying by 3, and also avoid `quarters_sub` to narrowing back
to i32 when calling addition after negative INT_MIN. Four date types
share this repair, and the real bias is transferred to the existing date
range for inspection and processing.

2. **FE quarter function calculation**: complete the `quarters_add` of
TIMESTAMP_NS, using long multiplication; four date types `quarters_sub`
directly calculate `-3L * quarter `, avoid losing in i32 first. The
original long addition of DATE/DATETIME/TIMESTAMPTZ and the return
mechanism after the failure of constant folding remain unchanged.

3. **Error message**: The offset parameter in the date cross-bound error
uses Int64 to avoid truncating the number of months again when the error
is reported.

### Release note

Fix quarters_add and quarters_sub returning incorrect dates for
excessively large quarter intervals; these inputs now report date-range
errors.

+++ be/src/exprs/function/function_date_or_datetime_computation.h
@@ -422,8 +422,9 @@ struct AddQuartersImpl {
 
     static constexpr auto name = "quarters_add";
     static constexpr auto is_nullable = false;
-    static inline ReturnValueType execute(const InputValueType& t, Int32 delta) {
-        return date_time_add<TimeUnit::MONTH, PType, Int32>(t, 3 * delta);
+    static inline ReturnValueType execute(const InputValueType& t, Int64 delta) {
+        // Preserve both the month offset and the negation of INT_MIN in quarters_sub.
+        return date_time_add<TimeUnit::MONTH, PType, Int64>(t, 3 * delta);
     }
 
     static DataTypes get_variadic_argument_types() {
